Invoke annotation processor manually #24
- Disable annotation processor in generator tests
- we can now compile enums with processor in a programmatic way
- we can, as a result of the above modification, run annotation
processor on multiple enums, testing different combination of
annotations.
